Refactoring repair on the client side
